# Conflicts:
#	services/CleaningService.java

+++ models/Apartment.java
@@ -1,7 +1,7 @@
 package models;
 
 public class Apartment extends Property {
-    public Apartment(double sizeInSquareMeters, double price, Address address,boolean isSold) {
+    public Apartment(int sizeInSquareMeters, int price, Address address,boolean isSold) {
         super(sizeInSquareMeters, price, address,isSold);
     }
     @Override
